+++ CHANGELOG.rst
@@ -7,6 +7,19 @@ foursight
 Change Log
 ----------
 
+
+4.9.0
+=====
+
+`PR 577: Get wf details from db not hard coding <https://github.com/4dn-dcic/foursight/pull/577>`_
+
+* Refactor of wfr_utils to utilize properties and tags on updated schemas for workflow and experiment_type to generate info about accepted versions of workflows and pipelines
+* Updated wfr_checks and wfr_encoded_checks to utilize the new info from db
+* provided additional keyword args to override the accepted versions (useful for testing or some cases where re-runs of updated pipelines are needed)
+* fix bug that allowed more than 2 errored fastq_first_line wfr runs
+* update poetry lock
+
+
 4.8.2
 =====